perf(api): optimize rate limiter middleware key formatting and concurrency
Replace fmt.Sprintf key construction with direct string concatenation and strconv.FormatInt in RateLimiterMiddleware to eliminate reflection overhead and heap allocations. Upgrade SafeCounter from sync.Mutex to atomic operations (atomic.AddInt64) for lock-free counter updates on the hot HTTP request path.

+++ back-end/cmd/api/middleware/rate_limiter.go
@@ -102,6 +102,6 @@ type SafeCounter struct {
 	v int64
 }
 
-func (c *SafeCounter) Inc() int64 {
-	return atomic.AddInt64(&c.v, 1)
+func (c *SafeCounter) Inc() int {
+	return int(atomic.AddInt64(&c.v, 1))
 }
